build(phpstan): lower the default level
The focus is on the testing and not on the static analysis. Whilst I'll
still use it locally and in CI as a quality tool, I don't think that it
needs to be set at the maximum level.

I'd prefer it if it didn't potentially didn't distract or confuse anyone
looking at the code who is trying to learn about testing and who may not
have been exposed to static analysis yet, and as the two can be done
independently of each others, it makes sense to me to lower the default
level set for PHPStan.